Moving constants declaration into dfovec.py
